build_dataset.py takes as input a dataset and the train/dev/test split ratio, performs Z-score normalization on it, and outputs a static train/dev/test set and the mean and standard deviation used to perform Z-score normalization.
nn_hyperpara_sceener.py and xgboost_hyperpara_screener.py allow to explore the space of possible hyperparameter configuratins for neural networks and xgboost. This can be done manually, by specifying the hyperparameters in a .csv config file, or automatically, by specifying the allowed ranges of the hyperparameters and employ the optimizer Optuna to tune the hyperparameters.